16-Observability/requirements.txt
fastapi[all]
uvicorn
gunicorn
pydantic
asyncpg
sqlalchemy
redis==4.3.3
fastapi-cache2[redis]==0.1.8
prometheus-fastapi-instrumentato==5.9.1
python-logging-loki==0.3.1